requirements.txt
arrow==0.12.1
astroid==1.6.4
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
BeautifulSoup==3.2.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
cycler==0.10.0
dateparser==0.7.0
docopt==0.6.2
enum34==1.1.10
et-xmlfile==1.0.1
funcsigs==1.0.2
futures==3.3.0
idna==2.6
isort==4.3.4
jdcal==1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
matplotlib==2.2.3
mccabe==0.6.1
moment==0.8.2
more-itertools==5.0.0
mpl-finance==0.10.0
numpy==1.14.1
openpyxl==2.5.3
pandas==0.23.4
pathlib2==2.3.3
PeakUtils==1.3.0
pluggy==0.8.1
py==1.7.0
pylint==1.9.1
pymongo==3.7.2
pyparsing==2.2.0
pytest==4.1.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==5.3
regex==2018.2.21
requests==2.21.0
scandir==1.9.0
scipy==1.2.0
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
times==0.7
tzlocal==1.5.1
urllib3==1.24.2
virtualenv==15.1.0
windows-curses==2.1.0
wrapt==1.10.11
xlrd==1.1.0
yagmail==0.11.224
